refactor utils to new file
