+++ ndc-http-schema/openapi/internal/oas2_schema.go
@@ -332,7 +332,7 @@ func (oc *oas2SchemaBuilder) getSchemaTypeFromProxy(schemaProxy *base.SchemaProx
 
 // Support converting allOf and anyOf to object types with merge strategy
 func (oc *oas2SchemaBuilder) buildUnionSchemaType(baseSchema *base.Schema, schemaProxies []*base.SchemaProxy, unionType oasUnionType, fieldPaths []string) (*SchemaInfoCache, error) {
-	proxies, mergedType, isNullable := evalSchemaProxiesSlice(schemaProxies, oc.location)
+	proxies, mergedType, isNullable, isEmptyObject := evalSchemaProxiesSlice(schemaProxies, oc.location)
 	nullable := isNullable || (baseSchema.Nullable != nil && *baseSchema.Nullable)
 	if mergedType != nil {
 		result, err := oc.getSchemaType(mergedType, fieldPaths)
@@ -348,6 +348,10 @@ func (oc *oas2SchemaBuilder) buildUnionSchemaType(baseSchema *base.Schema, schem
 
 	switch len(proxies) {
 	case 0:
+		if isEmptyObject {
+			return createSchemaInfoJSONScalar(oc.builder.schema, nullable), nil
+		}
+
 		oasTypes, isNullable := extractNullableFromOASTypes(baseSchema.Type)
 		if len(baseSchema.Type) > 1 || isPrimitiveScalar(baseSchema.Type) {
 			scalarName := getScalarFromType(oc.builder.schema, oasTypes, baseSchema.Format, baseSchema.Enum, fieldPaths)
@@ -401,6 +405,10 @@ func (oc *oas2SchemaBuilder) buildUnionSchemaType(baseSchema *base.Schema, schem
 			result.TypeSchema.Description = utils.StripHTMLTags(baseSchema.Description)
 		}
 
+		if isEmptyObject {
+			result = transformNullableObjectPropertiesSchema(oc.builder.schema, result, nullable, fieldPaths)
+		}
+
 		return result, nil
 	}
 
@@ -422,6 +430,10 @@ func (oc *oas2SchemaBuilder) buildUnionSchemaType(baseSchema *base.Schema, schem
 
 	result := mergeUnionTypeSchemas(oc.builder.schema, baseSchema, unionSchemas, unionType, fieldPaths)
 	result.OneOf = oneOfInfos
+
+	if isEmptyObject {
+		result = transformNullableObjectPropertiesSchema(oc.builder.schema, result, nullable, fieldPaths)
+	}
 
 	return result, nil
 }
